move ResourceContributor to @/components
