+++ apps/epic-web/src/app/(content)/[article]/page.tsx
@@ -5,9 +5,9 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { notFound } from 'next/navigation'
 import { ArticleCTA } from '@/app/(content)/[article]/_components/article-cta'
-import ResourceContributor from '@/app/(content)/[article]/_components/resource-contributor'
 import ContributorBio from '@/components/contributor-bio'
 import FloatingActionsBar from '@/components/floating-actions-bar'
+import ResourceContributor from '@/components/resource-contributor'
 import Share from '@/components/share'
 import { env } from '@/env.mjs'
 import { type Article } from '@/lib/articles'
