+++ apps/epic-web/src/app/(content)/articles/page.tsx
@@ -2,8 +2,8 @@ import * as React from 'react'
 import { Suspense } from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
-import ResourceContributor from '@/app/(content)/[article]/_components/resource-contributor'
 import FloatingActionsBar from '@/components/floating-actions-bar'
+import ResourceContributor from '@/components/resource-contributor'
 import { getArticles } from '@/lib/articles-query'
 import { getServerAuthSession } from '@/server/auth'
 import slugify from '@sindresorhus/slugify'
